Drop special normalization methods
This was a temporary solution, from now on only `field(normalize=...)`
is supported.
